+++ src/Encryptor.h
@@ -1,5 +1,7 @@
 ﻿#pragma once
 #include "shell/shell.h"
+#include <filesystem>
 
 // Global Vars
 const std::string VERSION = "v1.2";
+extern std::filesystem::path programFilePath;